Changes to getcsv

+++ HW6_vys217/getcsv.py
@@ -35,8 +35,8 @@ def getPLUTOcsv():
     '''
     print ("Downloading")
     ### First I will heck that it is not already there
-    if not os.path.isfile(os.getenv("PUIDATA") + "/" + "Manhattan/MNMapPLUTO.shp"):
-        if os.path.isfile("Manhattan/MNMapPLUTO.shp"):
+    if not os.path.isfile(os.getenv("PUIDATA") + "/" + "Manhattan/"):
+        if os.path.isfile("Manhattan/"):
             # if in the current dir just move it
             if os.system("mv " + "Manhattan " + os.getenv("PUIDATA")):
                 print ("Error moving file!, Please check!")
@@ -48,12 +48,13 @@ def getPLUTOcsv():
                 ###  To move it I use the os.system() functions to run bash commands with arguments
                 os.system("mv " + "mappluto_16v1.zip " + os.getenv("PUIDATA"))
             ### unzip the csv
-            os.system("unzip " + os.getenv("PUIDATA") + "/" + "mappluto_16v1.zip")
-            print("unzip " + os.getenv("PUIDATA") + "/" + "mappluto_16v1.zip")
-            os.system("rm -r Bronx/")
-            os.system("rm -r Bronx/")
-            os.system("rm -r Bronx/")
-            os.system("rm -r Bronx/")
+            if os.system("unzip " + os.getenv("PUIDATA") + "/" + "mappluto_16v1.zip"):
+                print("Unzipped")
+                os.system("mv " + "Manhattan " + os.getenv("PUIDATA"))
+                os.system("rm -r Bronx/")
+                os.system("rm -r Brooklyn/")
+                os.system("rm -r Queens/")
+                os.system("rm -r Staten_Island/")
     ### One final check:
     if not os.path.isfile(os.getenv("PUIDATA") + "/" + "Manhattan/MNMapPLUTO.shp"):
         print ("WARNING!!! something is wrong: the file is not there!")
